Need to convert value to pointer type

+++ aoc-05-02/src/main.rs
@@ -53,7 +53,7 @@ fn run_intcode(mut program: Vec<i32>, input: Option<i32>) -> (Vec<i32>, Vec<i32>
                 let test_value = get_value(&program, current_position, &current_inst, 0);
                 if test_value != 0 {
                     let jump_location = get_value(&program, current_position, &current_inst, 1);
-                    current_position = jump_location;
+                    current_position = jump_location as usize;
                 } else {
                     current_position += 3;
                 }
